remove checkpoints after loading

+++ core/solver.py
@@ -91,7 +91,8 @@ class Solver(nn.Module):
         if args.resume_iter > 0:
             self._load_checkpoint(args.resume_iter)            
             for name in os.listdir(args.checkpoint_dir):
-                os.remove(ospj(args.checkpoint_dir, name))
+                if name[:6] == f"{args.resume_iter:0>6}":
+                    os.remove(ospj(args.checkpoint_dir, name))
 
         # remember the initial value of ds weight
         initial_lambda_ds = args.lambda_ds
